feat(C++): implement 8

+++ cpp/registry.cpp
@@ -1,3 +1,3 @@
 #include "registry.hpp"
 
-std::map<int, std::function<int(void)>> solutions;
+std::map<int, std::function<int64_t(void)>> solutions;
