+++ cpp/solutions/solution_1.cpp
@@ -1,7 +1,7 @@
 #include "registry.hpp"
 
-int solution_1() {
-    int sum = 0;
+int64_t solution_1() {
+    int64_t sum = 0;
     for (int i = 0; i < 1000; ++i) {
         if (i % 3 == 0 || i % 5 == 0) {
             sum += i;
